Remove redundant cast conversions

+++ src/api/processors/rotation.cpp
@@ -39,8 +39,7 @@ VImage Rotation::process(const VImage &image) const {
             : image.bandjoin_const({255});  // Assumes images are always 8-bit
 
     return utils::stay_sequential(output_image, config_.process_timeout)
-        .rotate(static_cast<double>(rotation),
-                VImage::option()->set("background", background_rgba));
+        .rotate(rotation, VImage::option()->set("background", background_rgba));
 }
 
 }  // namespace weserv::api::processors
